Move CommitEditor into own source file for preparation to write tests
